Format crash method name in CrashDetector
Modified the crashData.getMethod().toString() execution to format the method name with the first letter capitalized and the rest of the name in lowercase. This change is implemented in the CrashDetector file to improve the readability of crash messages, ultimately enhancing the user experience during an event of a crash.