Implement ScareService and corresponding command permissions
A new ScareService class has been introduced to handle the 'scare' functionality, and a relevant permission command for the '/scarecrash' command has been added to the plugin.yml file. The ScareService class is responsible for creating a simulated virus injection experience to scare players. Relevant scare notifications as a part of this experience have also been moved from the PCCommand class into the new ScareService. The README file has been updated to reflect the addition of this permission.